dash-nginx-uwsgi-docker-master/requirements.txt
dash==1.0.0
dash-daq==0.1.0
Flask-Caching
pandas
dash-bootstrap-components
